Replace found cards with new cards in place
This, unfortunately, introduces something of a hack ala `blankCard` but
it works for now because `Card` is sufficiently secretive about its
inner workings.

There's only one place where we have to deal with it..and that's when we
look at a Card's properties to evaluate a set. Suffice to say, if a card
odesn't have properties (i.e. is `BlankCard`) then it cannot be a set.

`BlankCard` should not be selectable but that's not a huge issue at the
moment.